+++ docs/src/main.tsx
@@ -1,7 +1,6 @@
-import { $, component$, Host, useHostElement, useStore, useStyles$ } from '@builder.io/qwik';
-import { Builder } from './layouts/builder/builder';
-import { setHeadLinks, setHeadMeta, usePage } from '@builder.io/qwest';
+import { $, component$, Host, useStore, useStyles$ } from '@builder.io/qwik';
 import styles from './global.css';
+import { Page } from './components/page/page';
 
 export interface SiteStore {
   headerMenuOpen: boolean;
@@ -9,6 +8,7 @@ export interface SiteStore {
 }
 
 export const Main = component$(() => {
+  // setup
   useStyles$(styles);
 
   const store = useStore<SiteStore>({
@@ -16,30 +16,8 @@ export const Main = component$(() => {
     sideMenuOpen: false,
   });
 
-  return $(async () => {
-    const hostElm = useHostElement();
-    const page = await usePage(hostElm);
-
-    let body = <Builder store={store} />;
-
-    if (page) {
-      const attrs = page.attributes;
-      const Layout = page.layout;
-      const Content = page.content;
-
-      setHeadMeta(hostElm, {
-        title: attrs.title + ' - Qwik',
-        description: attrs.description,
-      });
-
-      setHeadLinks(hostElm, [{ rel: 'canonical', href: page.url.href }]);
-      body = (
-        <Layout store={store}>
-          <Content store={store} />
-        </Layout>
-      );
-    }
-
+  return $(() => {
+    // render
     return (
       <Host
         class={{
@@ -47,7 +25,7 @@ export const Main = component$(() => {
           'menu-open': store.sideMenuOpen,
         }}
       >
-        {body}
+        <Page store={store} />
       </Host>
     );
   });
