+++ starters/apps/starter-builder/src/components/footer/footer.tsx
@@ -1,7 +1,10 @@
 import { component$, $ } from '@builder.io/qwik';
 
 export const Footer = component$(() => {
+  // setup
+
   return $(() => (
+    // render
     <>
       <hr />
       <p style={{ 'text-align': 'center' }}>
