+++ starters/apps/starter-builder/src/components/header/header.tsx
@@ -1,7 +1,10 @@
 import { component$, $ } from '@builder.io/qwik';
 
 export const Header = component$(() => {
+  // setup
+
   return $(() => (
+    //render
     <p style={{ 'text-align': 'center' }}>
       <a href="https://github.com/builderio/qwik">
         <img
