+++ starters/apps/starter-partytown/src/main.tsx
@@ -2,8 +2,11 @@ import { component$, $, useEvent, useStore } from '@builder.io/qwik';
 import './global.css';
 
 export const Main = component$(() => {
+  // setup
   const state = useStore({ name: 'World', running: true });
+
   return $(() => {
+    // render
     return (
       <div id="my-app" onDocumentExpensiveComputationDone$={() => (state.running = false)}>
         <p style={{ 'text-align': 'center' }}>
