+++ starters/apps/starter/src/components/logo/logo.tsx
@@ -1,7 +1,10 @@
 import { $, component$, Host } from '@builder.io/qwik';
 
 export const Logo = component$(() => {
+  // setup
+
   return $(() => (
+    // render
     <Host style={{ 'text-align': 'center' }}>
       <a href="https://github.com/builderio/qwik">
         <img
