+++ starters/apps/todo/src/components/header/header.tsx
@@ -8,8 +8,11 @@ import type { Todos } from '../../state/state';
  */
 export const Header = component$(
   (props: { todos: Todos }) => {
+    // setup
     const state = useStore({ text: '' });
+
     return $(() => {
+      // render
       return (
         <>
           <h1>todos</h1>
