+++ starters/apps/todo/src/components/item/item.tsx
@@ -16,8 +16,11 @@ import type { TodoItem, Todos } from '../../state/state';
  */
 export const Item = component$(
   (props: { item: TodoItem; todos: Todos }) => {
+    // setup
     const state = useStore({ editing: false });
+
     return $(() => {
+      // render
       return (
         <Host class={{ completed: props.item.completed, editing: state.editing }}>
           <div class="view">
