+++ starters/apps/todo/src/main.tsx
@@ -15,6 +15,7 @@ import './index.css';
  * download to the client.
  */
 export const Main = component$(() => {
+  // setup
   const todos = useStore<Todos>({
     filter: 'all',
     items: [
@@ -23,7 +24,9 @@ export const Main = component$(() => {
       { completed: false, title: 'Profit' },
     ],
   });
+
   return $(() => {
+    // render
     return (
       <section class="todoapp">
         <Header todos={todos} />
